refactor: move message package form seperate other settings
